Bug fix from long to double

+++ AiPAM/src/main/java/com/jamdev/maven/aipam/layout/detectionDisplay/waveform/ScrollingPlotSegmenter.java
@@ -425,7 +425,7 @@ public abstract class ScrollingPlotSegmenter {
 		}
 		
 		ArrayList<PlotSegment> plotSegments = findPlotSegmentChunk(pamDataUnit.getTimeMilliseconds(), 
-				(long) (pamDataUnit.getTimeMilliseconds()+pamDataUnit.getDurationInMilliseconds())); 
+				 (pamDataUnit.getTimeMilliseconds()+pamDataUnit.getDurationInMilliseconds())); 
 		
 		//System.out.println("draw data unit: UID" + pamDataUnit.getUID() + " Image chunk: " + writableImageChunk);
